+++ tests/misc/tests/lamports/lamports.ts
@@ -20,4 +20,24 @@ describe(IDL.name, () => {
       .accounts({ signer, pda })
       .rpc();
   });
+
+  it("Can use the Lamports trait for checked arithmetic", async () => {
+    const signer = program.provider.publicKey!;
+    const [pda] = anchor.web3.PublicKey.findProgramAddressSync(
+      [Buffer.from("lamports")],
+      program.programId
+    );
+
+    await program.methods
+      .testLamportsTraitChecked(new anchor.BN(anchor.web3.LAMPORTS_PER_SOL))
+      .accounts({ signer, pda })
+      .rpc();
+  });
+
+  it("Throws when overflow occurs", async () => {
+    const signer = program.provider.publicKey!;
+    const [pda] = anchor.web3.PublicKey.findProgramAddressSync(
+      [Buffer.from("lamports")],
+      program.programId
+    );
 });